# server/docs/swagger.yaml
definitions:
...
  loomiadmin.ExchangePointsOrder:
    properties:
      ID:
        description: 主键ID
        type: integer
      accountName:
        description: 用户名
        type: string
      accountNumber:
        description: 银行账号、钱包、其他账户标识
        type: string
      addPoints:
        description: 变动的积分
        type: integer
      amount:
        description: 兑换金额
        type: string
      countryName:
        description: 国家
        type: string
      currency:
        description: 币种
        type: string
      desc:
        description: 描述
        type: string
      extInfo:
        description: 额外信息汇总；
        type: string
      mainChannel:
        description: 主渠道
        type: string
      merchantId:
        description: 商户id
        type: string
      operatorId:
        description: 审核人id
        type: integer
      orderNo:
        description: 商户订单号
        type: string
      pointsAfter:
        description: 修改后积分
        type: integer
      pointsBefore:
        description: 修改前积分余额
        type: integer
      pointsFlowId:
        description: 积分流水id
        type: integer
      productId:
        description: 产品id
        type: integer
      status:
        description: 对应订单状态 提现状态 1:待处理 2:待打款 10:打款失败 11:后台拒绝 21:成功
        type: integer
      subChannel:
        description: 子渠道; bank_code
        type: string
      uid:
        description: 用户uid
        type: string
    type: object
...
  request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.5.9
paths:
...
  /exchangePointsOrder/createExchangePointsOrder:
    post:
      consumes:
      - application/json
      parameters:
      - description: x-token
        in: header
        name: x-token
        type: string
      - description: 创建exchangePointsOrder表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/loomiadmin.ExchangePointsOrder'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"创建成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建exchangePointsOrder表
      tags:
      - ExchangePointsOrder
  /exchangePointsOrder/deleteExchangePointsOrder:
    delete:
      consumes:
      - application/json
      parameters:
      - description: x-token
        in: header
        name: x-token
        type: string
      - description: 删除exchangePointsOrder表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/loomiadmin.ExchangePointsOrder'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除exchangePointsOrder表
      tags:
      - ExchangePointsOrder
  /exchangePointsOrder/deleteExchangePointsOrderByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: x-token
        in: header
        name: x-token
        type: string
      - description: 批量删除exchangePointsOrder表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除exchangePointsOrder表
      tags:
      - ExchangePointsOrder
  /exchangePointsOrder/findExchangePointsOrder:
    get:
      consumes:
      - application/json
      parameters:
      - description: x-token
        in: header
        name: x-token
        type: string
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - description: 用户名
        in: query
        name: accountName
        type: string
      - description: 银行账号、钱包、其他账户标识
        in: query
        name: accountNumber
        type: string
      - description: 变动的积分
        in: query
        name: addPoints
        type: integer
      - description: 兑换金额
        in: query
        name: amount
        type: string
      - description: 国家
        in: query
        name: countryName
        type: string
      - description: 币种
        in: query
        name: currency
        type: string
      - description: 描述
        in: query
        name: desc
        type: string
      - description: 额外信息汇总；
        in: query
        name: extInfo
        type: string
      - description: 主渠道
        in: query
        name: mainChannel
        type: string
      - description: 商户id
        in: query
        name: merchantId
        type: string
      - description: 审核人id
        in: query
        name: operatorId
        type: integer
      - description: 商户订单号
        in: query
        name: orderNo
        type: string
      - description: 修改后积分
        in: query
        name: pointsAfter
        type: integer
      - description: 修改前积分余额
        in: query
        name: pointsBefore
        type: integer
      - description: 积分流水id
        in: query
        name: pointsFlowId
        type: integer
      - description: 产品id
        in: query
        name: productId
        type: integer
      - description: 对应订单状态 提现状态 1:待处理 2:待打款 10:打款失败 11:后台拒绝 21:成功
        in: query
        name: status
        type: integer
      - description: 子渠道; bank_code
        in: query
        name: subChannel
        type: string
      - description: 用户uid
        in: query
        name: uid
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询exchangePointsOrder表
      tags:
      - ExchangePointsOrder
  /exchangePointsOrder/getExchangePointsOrderList:
    get:
      consumes:
      - application/json
      parameters:
      - description: x-token
        in: header
        name: x-token
        type: string
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - description: 用户名
        in: query
        name: accountName
        type: string
      - description: 银行账号、钱包、其他账户标识
        in: query
        name: accountNumber
        type: string
      - description: 变动的积分
        in: query
        name: addPoints
        type: integer
      - description: 兑换金额
        in: query
        name: amount
        type: string
      - description: 国家
        in: query
        name: countryName
        type: string
      - description: 币种
        in: query
        name: currency
        type: string
      - description: 描述
        in: query
        name: desc
        type: string
      - in: query
        name: endCreatedAt
        type: string
      - description: 额外信息汇总；
        in: query
        name: extInfo
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 主渠道
        in: query
        name: mainChannel
        type: string
      - description: 商户id
        in: query
        name: merchantId
        type: string
      - description: 审核人id
        in: query
        name: operatorId
        type: integer
      - description: 商户订单号
        in: query
        name: orderNo
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - description: 修改后积分
        in: query
        name: pointsAfter
        type: integer
      - description: 修改前积分余额
        in: query
        name: pointsBefore
        type: integer
      - description: 积分流水id
        in: query
        name: pointsFlowId
        type: integer
      - description: 产品id
        in: query
        name: productId
        type: integer
      - in: query
        name: startCreatedAt
        type: string
      - description: 对应订单状态 提现状态 1:待处理 2:待打款 10:打款失败 11:后台拒绝 21:成功
        in: query
        name: status
        type: integer
      - description: 子渠道; bank_code
        in: query
        name: subChannel
        type: string
      - description: 用户uid
        in: query
        name: uid
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取exchangePointsOrder表列表
      tags:
      - ExchangePointsOrder
  /exchangePointsOrder/updateExchangePointsOrder:
    put:
      consumes:
      - application/json
      parameters:
      - description: x-token
        in: header
        name: x-token
        type: string
      - description: 更新exchangePointsOrder表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/loomiadmin.ExchangePointsOrder'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新exchangePointsOrder表
      tags:
      - ExchangePointsOrder
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"